requirements.txt
pyfiglet==0.8.post1
python-decouple==3.6
rich==12.0.1
termcolor==1.1.0
